refactor(shaders): split render.rs into focused submodules
Break the 860-line render.rs into render/mod.rs + 4 submodules:
- render/math.rs: normalize_vec3, cross, dot, compute_ray, pack_bgra, hash3, value_noise
- render/lighting.rs: compute_ao, march_shadow, apply_lighting, compute_lava_light, sun_direction
- render/materials.rs: textured_material_color, apply_emissive, emissive_lava/wood/gunpowder
- render/sky.rs: compute_sky_color, blackbody_color

Entry point (render_voxels) and shade_voxel stay in mod.rs.
Module path unchanged so SPIR-V entry point name is still
compute::render::render_voxels.

Closes #187
